triomphe

Struct ThinArc

pub struct ThinArc<H, T> { /* private fields */ }

A “thin” Arc containing dynamically sized data

This is functionally equivalent to Arc<(H, [T])>

When you create an Arc containing a dynamically sized type like HeaderSlice<H, [T]>, the Arc is represented on the stack as a “fat pointer”, where the length of the slice is stored alongside the Arc’s pointer. In some situations you may wish to have a thin pointer instead, perhaps for FFI compatibility or space efficiency.

Note that we use [T; 0] in order to have the right alignment for T.

ThinArc solves this by storing the length in the allocation itself, via HeaderSliceWithLength.

Implementations

impl<H, T> ThinArc<H, T>

pub fn with_arc<F, U>(&self, f: F) -> U
where F: FnOnce(&Arc<HeaderSlice<HeaderWithLength<H>, [T]>>) -> U,

Temporarily converts |self| into a bonafide Arc and exposes it to the provided callback. The refcount is not modified.

pub fn with_arc_mut<F, U>(&mut self, f: F) -> U
where F: FnOnce(&mut Arc<HeaderSlice<HeaderWithLength<H>, [T]>>) -> U,

Temporarily converts |self| into a bonafide Arc and exposes it to the provided callback. The refcount is not modified.

pub fn from_header_and_iter<I>(header: H, items: I) -> Self
where I: Iterator<Item = T> + ExactSizeIterator,

Creates a ThinArc for a HeaderSlice using the given header struct and iterator to generate the slice.

pub fn from_header_and_slice(header: H, items: &[T]) -> Self
where T: Copy,

Creates a ThinArc for a HeaderSlice using the given header struct and a slice to copy.

pub fn ptr(&self) -> *const c_void

Returns the address on the heap of the ThinArc itself – not the T within it – for memory reporting.

pub fn heap_ptr(&self) -> *const c_void

Returns the address on the heap of the Arc itself – not the T within it – for memory reporting.

pub unsafe fn from_raw(ptr: *const c_void) -> Self

Safety

Constructs an ThinArc from a raw pointer.

The raw pointer must have been previously returned by a call to ThinArc::into_raw.

The user of from_raw has to make sure a specific value of T is only dropped once.

This function is unsafe because improper use may lead to memory unsafety, even if the returned ThinArc is never accessed.

pub fn into_raw(self) -> *const c_void

Consume ThinArc and returned the wrapped pointer.

pub fn as_ptr(&self) -> *const c_void

Provides a raw pointer to the data. The counts are not affected in any way and the ThinArc is not consumed. The pointer is valid for as long as there are strong counts in the ThinArc.

pub fn strong_count(this: &Self) -> usize

The reference count of this Arc.

The number does not include borrowed pointers, or temporary Arc pointers created with functions like ArcBorrow::with_arc.

The function is called strong_count to mirror std::sync::Arc::strong_count, however triomphe::Arc does not support weak references.